refactor: move Statistic into User
